requirements.txt
fastapi
pydantic
uvicorn
tensorflow==2.12.0
pandas
numpy
networkx
osmnx
python-dotenv
pyyaml
python-multipart